pyproject.toml
[tool.poetry]
name = "labs"
version = "0.1.0"
description = ""
authors = [
    "Carlos Martinez <c.martinez@runtime-revolution.com>",
    "Filipa Silva <filipa.silva@runtime-revolution.com>",
    "Luis Moreira <l.moreira@runtime-revolution.com>",
    "Sérgio Ildefonso <s.ildefonso@runtime-revolution.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.7"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
transformers = "^4.42.4"
nltk = "^3.8.1"
spacy = "^3.7.5"
spacytextblob = "^4.0.0"
langchain = "^0.2.10"
litellm = "^1.41.25"
gitpython = "^3.1.43"
pyicu = "^2.13.1"
pycld2 = "^0.41"
langchain-community = "^0.2.9"
python-json-logger = "^2.0.7"
pathspec = "^0.12.1"
pgvector = "^0.3.2"
pytest-postgresql = "^6.0.0"
docker = "^7.1.0"
fastapi = "^0.111.1"
graphviz = "^0.20.3"
pytest = "^8.3.2"
rouge-score = "^0.1.2"
pandas = "^2.2.2"
langdetect = "^1.0.9"
ruff = "^0.6.2"
pre-commit = "^3.8.0"
torch = "2.2.2" # Locked to this version because 
psycopg2-binary = "^2.9.9"
psycopg = {extras = ["binary"], version = "^3.2.2"}

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
jupyterlab = "^4.2.4"
vcrpy = "^6.0.1"
pytest = "^8.3.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"